Implement first semantic rewrite: ExplicitImplicit.
This commit implements a semantic rewrite by using a compiler plugin.
The rewrite is executed during compilation (at least for now) if the
compiler plugin is loaded.

The sbt plugin injects the compiler plugin only for the
`scalafix` command, otherwise the compiler plugins does not
mess with any library dependencies.