Perform traversal of full type-term lookup tree
Given the ambiguity of term selects in Scala meta, we have to try luck
with every name and look it up with both type and term names. The
previous traversal assumed that once we found a symbol, we took the
right path, but this is not true:

```scala
object A
class A {
  type B
}
```

Looking up `A.B` would fail if the algorithm start looking term names.
`object A` would be found, but `type B` wouldn't since it's not defined
in the companion object but the type.

The new algorithm allow us to backtrack once we find these lookup
failures (as we're browsing through the RootPackage members, we know
that we will find at least one solution). Therefore, we perform a full
DFS pre-order traversal in the binary tree that represents all the
possibilities.