Track down performance regression in integration tests.
Doing .symbol on every Ref in a Traverser is apparently really slow.
This provides a valuable lesson: scalafix should not run rewrites during
compilation, scalafix should run rewrites against a semanticdb.
Why? By dumping the semanticdb first, we compile as fast as possible AND
it allows us to run rewrites on multiple compilation units in parallel.