Reorganize SBT tests (#88)
* Reorganize SBT tests

The SBT tests contain some duplication. This commit improves the situation,
by using symbolic links to a template project. Note that this approach is
heavily inspired by ensime-sbt (see https://github.com/ensime/ensime-sbt)
that uses a similar trick.

Thanks to this technique, `ScalafixTestUtility` is no longer published with
the plugin, but is visible to tests only. Also, we are now free to add
additional dependencies to our SBT tests (like `java-diff-utils`) without
interfering with the dependencies of the plugin.

* Improved error messages in SBT tests

This commit makes use of the recent reorganization of the SBT tests by using
`java-diff-utils` to generate better error messages.