Make scalafix full cross-version.
Scalafix depends on scala-compiler/reflect both for runtime compilation of
rewrites and scalahost's online mirror. Different versions of
scala-compiler and scala-reflect should not be mixed on the same
classpath, even for patch versions. However, we'll try to only publish
scalafix for the latest patch version to simplify the release step.
I'm happy to reconsider publishing scalafix for multiple patch version,
if there is demand for it