Clean up cli.
This previous implementation was quite a mess. This changes the cli
to more properly aggregate and propagate error handling. This also made
it easier to adapt to how the Semantic API requires the scala.meta.Input
to be reused. Previously, scalafix created its own Input but now it
re-uses the Input from scala.meta.Database.