Upgrade (#163)
* Upgrade to scalameta 1.8

* Fix failing tests

* Clean up cli.

This previous implementation was quite a mess. This changes the cli
to more properly aggregate and propagate error handling. This also made
it easier to adapt to how the Semantic API requires the scala.meta.Input
to be reused. Previously, scalafix created its own Input but now it
re-uses the Input from scala.meta.Database.

* Reorganize

* Disable cache.

It took 5 minutes to load the cache and still sbt downloaded everything
from scratch

* Fix last few bugs in cli.

- Skip compiling rewrites if inputs.isEmpty.
- Print warning if inputs is empty.
- More cleanup

* Upgrade platform docker image in CI.

* s/--debug/--verbose/