Lazily compute mirror so auto-detect --syntactic.
This commit refactors how mirrors are loaded in the cli. Before, mirrors
were eagerly loaded depending on whether --syntactic was passed or not.
This meant is was annoying to run syntactic rewrites, or you loaded
the semanticdb even if you didn't need it. After this commit, the cli
will only try to load the mirror if it's requested by a semantic
rewrite. This fix works for default rewrites along with
classloaded/compiled rewrites.

The most important change in this commit was to change `Option[Mirror]`
into `type LazyMirror = RewriteKind => Option[Mirror]`. This allowed
us to delay the computation of the mirror up to the point where we
needed to load the first semantic rewrite.

Along the way, I had to refactor a lot of things on the side, which
caused the large diff. The side changes are mostly

- Consolidate how cli arg --rewrites and .scalafix.conf rewrites are loaded
- Remove --syntactic since it's no longer used.
- Use Database instead of Mirror everywhere.
- Refactor validation logic inside CliRunner.Builder.
- More polished --verbose logging statements.