Remove unused imports. (#239)
* Remove unused imports.

I've been delaying this commit for too long hoping to fix it with
scalafix. However, it turns out that we can't yet use scalahost in the
scalafix repo because of https://github.com/scalameta/scalameta/issues/979.

I could not get myself to manually remove the imports so I implemented
instead an sbt plugin called sbt-messagehost that collects reported messages
from the compilerReporter in sbt and writes .semanticdb files. Using
messagehost, I was able to run RemoveUnusedImports on the scalafix
codebase. I suspect messagehost may become popular as a lightweight
alternative to scalahost.

I plan to publish messagehost later, it will not be part of the scalafix
repo.

* Format.