Implement crude MoveSymbol patch.
This patch can be used to move or rename call-sites of a global
definition like a class, trait, object or static def. This
implementation is crude, it is likely produce invalid code in many
cases. The patch tries to preserve the prefix when possible. For
example:

```
// config
from = a.b.c
to = d.b.x
// code before
import a.b.c
import a.b
b.c
c
// code after
import a.b
import d.b
import d.b.x
b.x
x
```

Room for improvement:

- Observe that `import a.b` is not removed as it may be used to prefix
  other identifiers from the `a.b.` package. It should be possible to
  remove the imports if its not used elsewhere.
- It's possible that duplicate imports are added, for example if
  the source contains `import a.b_` then scalafix may still add
  redundant `import a.b.c`.
- Imports are always inserted fully qualified at the bottom of the
  global imports list. It might be nice if it handled
  - imports grouped by curly braces
  - import sorting
  - local imports