Remove dependency on scalameta/paradise.
scalameta/paradise will not upgrade to scalameta v2.  This commit
removes the usage of the `@DeriveConfDecoder` macro annotation in favor
of manual decoders. Metaconfig is barely holding its weight at this
point, at it's a question whether we should remove it entirely in favor
of an established library like Circe. The code interfacing with
metaconfig is quite small, so maybe it's possible.