Fix #255, fallback to default config in ExplicitReturnTypes.
I find it a bit unsatisfying that rewrites must throw exceptions to
communicate errors in the configuration. Moreover, it's unnecessary
that the rewrite must validate the configuration on every RewriteCtx
instead of only once during construction. This fix is a short term
solution, we should eventually try to solve the root of the issue.