Generalize to DottyKeywords.
DottyKeywords backticks identifiers which are upcoming keywords in
Dotty. Note that DottyKeywords will rename `@inline` to

```scala
@`inline` def foo
```

This compiles with both scalac and dotc, but looks weird. In the future,
we may want the rewrite to be semantic and insert `inline` for
codebases targeting only Dotty and keep `@scala.inline`
(unbackticked) for codebases that want to target both compilers.