Move scalafix.config to scalafix.internal.config.
Since we promise binary compatibility from 0.5.0 onwards, I'm using this
last chance to reorganize things. The configuration case classes break
binary compatilibity with every new field in the `.copy` method.
It will be safe to access fields in the case classes, but not construct
them with either <init> or apply.