Add a few deprecated forwarders.
It's a bit awkward that ScalafixConfig is exposed in the public Rewrite
api while it's private. Hand-writing the `withX` methods for all nested
fields is a major pain. The getters will remain binary compatible so I
think it's fine that users access them even if the class is inside the
internal package.