Add test for deprecated names, VolatileLazyVal => DottyVolatileLazyVal.
I've been meaning to rename this rewrite for a while but wasn't
satisfied that this would break people's configuraiton. I wanted
to solve this potentially common problem in the future more elegantly.

Now that rewrites can have multiple names and names can have deprecation
warnings, this commits tests that it indeed works as expected.