Reorganize reporters.
- cliArg for reporting errors on cli args.
- diagnostic for regular diagnosticts
- lint.reporter for linter errors.
- reporter.reset to start fresh reporter.
- ScalafixConfig.withFreshReporter

I'm starting to doubt if using mutable state inside the reporter API was
a good idea. The right thing to do is to refactor the codebase to be
referentially transparent. However, I don't think metaconfig is a good
enough library to support that style of error handling. For example,
it was so much simpler to report the "Is --sourceroot correct?" warnings
via the reporter instead of through `Configured[T]`.