Add a sanity test for unicode characters.
This test checks that scalafix won't mix up unicode and non-unicode
identifiers. This case appeared in akka, where I thought scalafix had
inserted a cyclic dependency but it was in fact caused by scalariform
writing turning the identifierr token -> in type position into
it's unicode equivalent. This is a bug in scalariform, since -> and
it's uniocde characters are only treated equally for non-ident tokens.