handle coursier's --standalone classloader
When coursier created a standalone executable for Scalafix, then the app
is run in an unusual classloader. The thing is a URLClassLoader, but
urls of the individual jars are  weird - their getProtocol is not file,
but bootstrapsomethingsomething, and their getPath doesn't point to
anything on the file system.

This is not at all a problem for running Scalafix, but it becomes an
issue when Scalafix is doing runtime compilation to load rules using
the file: protocol. The Scala compiler instance that's spun up for
runtime compilation doesn't understand those weird urls, so compilation
fails.

The proposed patch fixes the compilation failure by extracting the jars
into temporary files on disk, and then passing paths to those files to
the Scala compiler instance.