Add scalafixEnable command.
This command should work for almost any build without fiddling with
scalacOptions or libraryDependencies. The problem is that scalafix
requires these keys to be correct

1. scalaVersion
2. scalacOptions
3. libraryDependencies

If the build configures `scalacOptions := Nil` then there is no way for
sbt-scalafix to automatically work out of the box.

For build where the automatic "addSbtPlugin => `> scalafix`"
installation instructions that don't work we can now recommend
`scalafixEnable` as simple explicit installation step.  In 0.6.0, we
should remove the automatic installation of scalacOptions and
libraryDependencies to have a single installationn instruction that
works for all builds.

This commit reorganizes the code a bit in ScalafixPlugin to group
together relevant settings.