Refactor LintMessage & Check for typos when enabling rules
I was writing a test to check if a scalafix:on for a rules foo has
a matching scalafix:off.

For example we can have a typo in scalafix:on that prevents the rule for
being enabled back again.

However in the test the owner was appended to the LintMessage returned
by the unmatched scalafix:on.

for example:

/*
rules = [
  Disable
  NoInfer
]

Disable.symbols = [
  "scala.Option.get"
]

NoInfer.symbols = [
  "scala.Predef.any2stringadd"
]
*/

package test.escapeHatch

object EscapeHatchMultipleRules {

  // scalafix:off NoInfer.any2stringadd, Disable.get
  1 + "foo"

  val a: Option[Int] = Some(1)
  a.get

   /* scalafix:on NoInfer.any2stringadd, Disable.gets */ // assert: UnusedScalafixOn
}

was returning Disable.UnusedScalafixOn and  and NoInfer.UnusedScalafixOn