Simplify ci-slow.
The CI failed while publishing v0.5.4, I'm not able to reproduce the
compile error in cli/doc so I suspect the problem is related to the
hairy ci-slow command, which triggered commands from the `it:test` task.
This commit removes that part and uses sbt more idiomatically, making
ci-slow stay within the task graph.