Strip trailing whitespace from testkit tests.
The formatting of the testkit messages leaks trailing whitespace,
his commit strips it from the test instead of fixing the formatting
f the messages to remove the creation of that whitespace in the
irst place. This is only for testkit messages so it's not critical
o have no trailing whitespace.