Fix #493, handle synthetics and symbol signatures in Disable.
Previously, Disable reported an error on the positions of any
ResolvedName to a disabled symbol, including Input.Synthetic and
Input.Denotation which are invisible and don't appear in the source
file. Now Disable ignores symbol signatures and for synthetics it
points a caret at the position where the synthetic got expanded
and includes the pretty-printed representation of the synthetic
to help make sense of the error message.