Cache jars and classloaders in sbt-scalafix.
There were reported heap-space usage problems that are potentially
caused by leaking classloaders in the sbt plugin. I thought that the sbt
task engine cached task results if the depended on settings/tasks don't
change.  This commit is a quick fix to test the hypothesis whether the
sbt plugin is causing problems