Move MissingFinal, Disable, ExplicitResultTypes and dotty rules out of the core repo
This does not mean those rules become more difficult to use, those rules
can still be automatically included by sbt-scalafix and tab completed
like builtin rules.

The primary purpose of this change is to enable rules to evolve faster.
The day-to-day bug fixes and improvements can be managed in smaller
repos with faster-running test suites. The rules will not be married
to the slower Scalafix release cycle.

This change is motivated by our experience with scala-collection-compat.
The collection migration rewrites have been able to evolve and improve
outside of the Scalafix core repo at a much faster pace than then would
have in the core repo.