Polish handling of trailing commas for removed imports
This was awfully painful since the code is tricky to follow. It would
probably be significantly simpler if we simply removed all import tokens
and synthesized tokens for remaining comments. Either way, the primary
improvement from this commit is that we have a slightly simpler rule for
deciding how to remove commas

- always remove leading comma (for importers + importees)
- remove trailing comma for the last unused importer/importee before
  the first used importer/importee.