Implement holistic pretty-printing strategy for scalafix.v1 types
- `toString` pretty-prints psuedo syntax, trying where possible to mimic
  actual code while still printing synthetic modifiers like `<default>`.
- `structure` pretty-prints the raw structure, showing class names.
- `structureLabeled` pretty-prints the raw structure, showing class and field names.
  Field names are particularly helpful in scalafix.v1 since types like
  STree and SType don't expose `unapply` for pattern matching.

This commit introduces a fixture to run "expect tests" to assert that
the pretty-printer does not regress in future releases. Expect tests
have been very helpful for us in scalameta so I'm excited to bring them
to scalafix.

It's a shame that we can't use case classes, but they are unfortunately
unusable if we ever add a new field to the public types. This
boilerplate is cumbersome to write but it allows us to evolve without
binary breaking changes.

All pretty-printing methods implement line-wrapping out of the box
thanks to the excellent Paiges library.

This PR upgrades to the latest Scalameta milestone and aligns the field
names that got renamed in https://github.com/scalameta/scalameta/pull/1760.