Use symbolic names for DocConstants.
It should at least be consistent, before it was a mix.  The downside to
this is that backticked wildcard `_` is not a valid identifier.  In this
case I think it's fine, whereas I'm convinced that
`scala.meta.Token.{KwNew,LeftParen}` was a great call. Type names should
not be symbolic.