Add more built-in Rule methods for documentation purposes
This commit adds several optional methods to `Rule` that are helpful
for documenting the purpose of each rule. This information can be used
for example by the sbt plugin to report an error when running a semantic
rule without having enabled SemanticDB.

Previously, the rule descriptions started with "Rewrite that..." or
"Linter that...". Now rules can document instead with `override def
isRewrite = true` whether they do rewriting or not.