Rename SType and STree to ScalaType and SyntheticTree
Based on feedback from my Scalafix workshop at Scala Italy, I think
these names are more self-documentary. Additionally, this commit
turns ScalaType and SyntheticTree into standard case classes.
This means we will need to do a breaking release in case one of the
types adds a new field, but I believe this change is worth it given
the improved readability of the source code and ability to use unapply
for pattern matching.