Only remove large lower bounds from large existential types.
Previously, we removed the lower and upper bound in large existential
types resulting in code that no longer compiler. Now, we only remove the
lower bound and keep the upper bound to keep the code compiling. Also,
previously, we didn't shorten quantifiers in existential types but
now we shorten them like other types.