Respect the intent of rules authors when returning atomic patches
In scalafixFileEvaluation, there is a possibility to apply only some patches.
The current issue, is that we were returning the low level patches, by recursively decomposing even Atomic patches.

If the author rule has created an atomic patch that contains two Add patches, we should not decompose it,
and therefore `previewPatches()` can only be called with "atomic" patches