enable imports for path-dependent types
Remove related TODO as we do want to favor `NestedObj.T4` reference
rather than `T4` when the parent is an object, or `Obj.NestedObj.T4`
if the grand parent is also an object.

For example, we want Term.Name (with import scala.meta.Term) rather
than Name (with import scala.meta.Term.Name).