Refactoring support for `given` imports (#203)
This PR makes a bunch of refactoring changes to simplify the implementation, as well as fixing one bug: when sorting importees, the `SymbolsFirst` option can be violated when `given` imports are there. For example:

```scala
import foo.{given, _}
```

The above import should be rewritten into

```scala
import foo.{_, given}
```

when `importSelectorsOrder` order is set to `SymbolsFirst`.