force usage of shaded TPrint
Injecting instances of pprint.TPrint to customize the docs no longer
has effect as of ffcd8b9a, which was a silent breaking change. This
commit exposes the breaking change so that rules built against the
latest scalafix-core either drop usage of TPrint or switch to the
shaded version which is used internally by metaconfig.