better error for external rules ran with old scalafix-cli version
metaconfig-pprint brought by metaconfig-core 0.10.0 will not be
available to external rules when classloaded by an old scalafix-cli
version as the runtime scalafix-core and its dependencies are driven by
scalafix-cli dependencies, not the rule dependencies.

We obviously have no control over old scalafix-cli versions, but since
the reference to the potentially-missing type is from a macro-generated
block, this wraps it to provide a more actionable error message.