RuleDecoder: avoid error on Scala 3 & stale class on Scala 2
In scala 3's AbstractFileClassLoader, `loadClass` is routing loadClass
calls to findClass, causing "attempted duplicate class definition"
when loading a rule sharing the same FQN as a previously compiled one.

In scala 2, there is no failure but loaded rules could be the result
of previous compilations, so returning a fresh classloader each time
is probably what we want.