Add Sonatype statistics
Fetch statistics from the Sonatype API and write the results in the repository as CSV files. Also show the data as bar charts in an HTML page available at `sonatype-stats/index.html` (it can directly be opened in a web browser, no need to set up an HTTP server).

Set up a monthly GitHub workflow that updates the stats.