OrganizeImports: document confusing behavior with star symbol
Even when a * symbol exists, Scala 3 (and the Scalameta parser with the
Scala 3 dialect) will treat imports with that syntax as wildcards. For projects
cross-building in Scala 2 and 3 and explicitly referencing such a symbol in the
importers, runs of OrganizeImports forcing Scala 2 syntax will yield
inconsistent results, as * would confusingly (but not so incorrectly) be
downgraded to _ in the Scala 3 run, would it will remain a reference to the
symbol in the Scala 2 run.

I attempted to align the Scala 3 run behavior with the Scala 2 run one, but
after a while it became clear that it would require a big redesign of the flow
(because the original syntax behind the wildcard gets lost across tree copies)
and would add a lot of complexity for very little value.

Also, I believe the current behavior is acceptable as the Scala 3 run output is
stable against a Scala 2 run. That is why I decided to simply document the
behavior via a test (which should hopefully prevent changes that would remove
that property in the future) instead of fixing it.