forward compatibility for cli full Scala version lookups
An old scalafix-interfaces client should be able to get the full Scala version
for a recent scalafix-cli version. This is achieved by encapsulating property
file lookup behind a Java interface, for which an implementation can easily
be retrieved via ServiceLoader, as long as the recent scalafix-versions JAR
has been fetched.